refactor: new group named AddressCard and JuiceboxCard
