+++ components/ActionLabel/TransferActionLabel.tsx
@@ -1,7 +1,7 @@
 import { getContractLabel } from "@/constants/Contract";
 import { numToPrettyString } from "@/utils/functions/NumberFormatter";
 import { Transfer } from "@/models/NanceTypes";
-import FormattedAddress from "@/components/Address/FormattedAddress";
+import FormattedAddress from "@/components/AddressCard/FormattedAddress";
 
 export default function TransferActionLabel({
   transfer,
