+++ components/JuiceboxCard/ResolvedProject.tsx
